requirements.txt
pyqt6
capstone
pyyaml
pytest